# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop API",
        "description": "API documentation for Shop application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация нового пользователя",
                "description": "Регистрирует нового пользователя в системе",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная регистрация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/UserResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories",
                "description": "Get list of categories",
                "operationId": "e225c2b7eb5daf7fb16e00f4f07ff030",
                "parameters": [
                    {
                        "name": "include_children",
                        "in": "query",
                        "description": "Include children categories in response",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "show_hidden",
                        "in": "query",
                        "description": "Show hidden categories",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{slug}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category by slug",
                "description": "Get category by slug",
                "operationId": "6b5e99ab9669011f1260b2a8fb93392e",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryNotFound"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "RegisterRequest": {
                "required": [
                    "first_name",
                    "last_name",
                    "middle_name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "first_name": {
                        "description": "Имя",
                        "type": "string",
                        "example": "Иван"
                    },
                    "last_name": {
                        "description": "Фамилия",
                        "type": "string",
                        "example": "Иванов"
                    },
                    "middle_name": {
                        "description": "Отчество",
                        "type": "string",
                        "example": "Иванович"
                    },
                    "email": {
                        "description": "Email",
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "password": {
                        "description": "Пароль",
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    },
                    "password_confirmation": {
                        "description": "Подтверждение пароля",
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    },
                    "phone": {
                        "description": "Телефон",
                        "type": "string",
                        "example": "+7 999 123 45 67"
                    },
                    "country": {
                        "description": "Страна",
                        "type": "string",
                        "example": "Россия"
                    },
                    "region": {
                        "description": "Регион",
                        "type": "string",
                        "example": "Московская область"
                    },
                    "city": {
                        "description": "Город",
                        "type": "string",
                        "example": "Москва"
                    },
                    "postal_code": {
                        "description": "Почтовый индекс",
                        "type": "string",
                        "example": "123456"
                    },
                    "address": {
                        "description": "Адрес",
                        "type": "string",
                        "example": "ул. Пушкина, д. 1"
                    }
                },
                "type": "object"
            },
            "Category": {
                "required": [
                    "id",
                    "name",
                    "slug"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Все свечи"
                    },
                    "slug": {
                        "type": "string",
                        "example": "vse-svechi"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание категории",
                        "nullable": true
                    },
                    "icon": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/1/icon.jpg",
                        "nullable": true
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 0
                    },
                    "is_visible": {
                        "type": "boolean",
                        "example": true
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CategoryCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object"
            },
            "CategoryRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Новая категория"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание новой категории",
                        "nullable": true
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 0
                    },
                    "icon": {
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CategoryNotFound": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Запрашиваемый ресурс не найден"
                    }
                },
                "type": "object"
            },
            "ApiResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Операция выполнена успешно"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The email field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Иван"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Иванов"
                    },
                    "middle_name": {
                        "type": "string",
                        "example": "Иванович"
                    },
                    "full_name": {
                        "type": "string",
                        "example": "Иванов Иван Иванович"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+7 999 123 45 67"
                    },
                    "address": {
                        "properties": {
                            "country": {
                                "type": "string",
                                "example": "Россия"
                            },
                            "region": {
                                "type": "string",
                                "example": "Московская область"
                            },
                            "city": {
                                "type": "string",
                                "example": "Москва"
                            },
                            "postal_code": {
                                "type": "string",
                                "example": "123456"
                            },
                            "address": {
                                "type": "string",
                                "example": "ул. Пушкина, д. 1"
                            }
                        },
                        "type": "object"
                    },
                    "email_verified": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "tags": [
        {
            "name": "Categories",
            "description": "API Endpoints for categories"
        },
        {
            "name": "Auth",
            "description": "Auth"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ]
}